Refactor tv show remote data source module structure; update package names and file locations, add network error handling utilities
