--- datasource/remote/mediaDetails/src/main/java/com/datasource/remote/mediadetails/models/tvShow/TvShowAuthorDetailsDto.kt
@@ -1,13 +1,0 @@
-package com.datasource.remote.mediadetails.models.tvShow
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TvShowAuthorDetailsDto(
-    @SerialName("name") val name: String? = null,
-    @SerialName("username") val username: String? = null,
-    @SerialName("avatar_path") val avatarPath: String? = null,
-    @SerialName("rating") val rating: Double? = null
-)
-
