--- datasource/remote/mediaDetails/src/main/java/com/datasource/remote/mediadetails/models/tvShow/TvShowGenreDto.kt
@@ -1,10 +1,0 @@
-package com.datasource.remote.mediadetails.models.tvShow
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TvShowGenreDto(
-    @SerialName("id") val id: Int? = null,
-    @SerialName("name") val name: String? = null
-)
