--- datasource/remote/mediaDetails/src/main/java/com/datasource/remote/mediadetails/models/tvShow/TvShowProductionCompanyDto.kt
@@ -1,12 +1,0 @@
-package com.datasource.remote.mediadetails.models.tvShow
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TvShowProductionCompanyDto(
-    @SerialName("id") val id: Int? = null,
-    @SerialName("logo_path") val logoPath: String? = null,
-    @SerialName("name") val name: String? = null,
-    @SerialName("origin_country") val originCountry: String? = null
-)
