--- datasource/remote/mediaDetails/src/main/java/com/datasource/remote/mediadetails/models/tvShow/TvShowProductionCountryDto.kt
@@ -1,10 +1,0 @@
-package com.datasource.remote.mediadetails.models.tvShow
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TvShowProductionCountryDto(
-    @SerialName("iso_3166_1") val iso31661: String? = null,
-    @SerialName("name") val name: String? = null
-)
