--- datasource/remote/mediaDetails/src/main/java/com/datasource/remote/mediadetails/models/tvShow/TvShowSimilarDto.kt
@@ -1,22 +1,0 @@
-package com.datasource.remote.mediadetails.models.tvShow
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TvShowSimilarDto(
-    @SerialName("adult") val adult: Boolean? = null,
-    @SerialName("backdrop_path") val backdropPath: String? = null,
-    @SerialName("genre_ids") val genreIds: List<Int>? = null,
-    @SerialName("id") val id: Int? = null,
-    @SerialName("original_language") val originalLanguage: String? = null,
-    @SerialName("original_title") val originalTitle: String? = null,
-    @SerialName("overview") val overview: String? = null,
-    @SerialName("popularity") val popularity: Double? = null,
-    @SerialName("poster_path") val posterPath: String? = null,
-    @SerialName("release_date") val releaseDate: String? = null,
-    @SerialName("title") val title: String? = null,
-    @SerialName("video") val video: Boolean? = null,
-    @SerialName("vote_average") val voteAverage: Double? = null,
-    @SerialName("vote_count") val voteCount: Int? = null
-)
