+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowAuthorDetailsDto.kt
@@ -1,0 +1,13 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowAuthorDetailsDto(
+    @SerialName("name") val name: String? = null,
+    @SerialName("username") val username: String? = null,
+    @SerialName("avatar_path") val avatarPath: String? = null,
+    @SerialName("rating") val rating: Double? = null
+)
+
