+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowCastDto.kt
@@ -1,0 +1,13 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowCastDto(
+    @SerialName("id") val id: Int? = null,
+    @SerialName("name") val name: String? = null,
+    @SerialName("profile_path") val profilePath: String? = null,
+    @SerialName("character") val character: String? = null
+)
+
