+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowCreditsDto.kt
@@ -1,0 +1,12 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowCreditsDto(
+    @SerialName("id") val id: Int? = null,
+    @SerialName("cast") val cast: List<TvShowCastDto>? = null,
+    @SerialName("crew") val crew: List<TvShowCrewDto>? = null
+)
+
