+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowGenreDto.kt
@@ -1,0 +1,10 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowGenreDto(
+    @SerialName("id") val id: Int? = null,
+    @SerialName("name") val name: String? = null
+)
