+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowImagesDto.kt
@@ -1,0 +1,12 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowImagesDto(
+    @SerialName("backdrops") val tvShowBackdropDto: List<TvShowBackdropDto>? = null,
+    @SerialName("id") val id: Int? = null,
+    @SerialName("logos") val logos: List<TvShowLogoDto>? = null,
+    @SerialName("posters") val posters: List<TvShowPosterDto>? = null
+)
