+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowNetworkDto.kt
@@ -1,0 +1,12 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowNetworkDto(
+    @SerialName("id") val id: Int? = null,
+    @SerialName("logo_path") val logoPath: String? = null,
+    @SerialName("name") val name: String? = null,
+    @SerialName("origin_country") val originCountry: String? = null
+)
