+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowProductionCountryDto.kt
@@ -1,0 +1,10 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowProductionCountryDto(
+    @SerialName("iso_3166_1") val iso31661: String? = null,
+    @SerialName("name") val name: String? = null
+)
