+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowReviewDto.kt
@@ -1,0 +1,16 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowReviewDto(
+    @SerialName("author") val author: String? = null,
+    @SerialName("author_details") val authorDetails: TvShowAuthorDetailsDto? = null,
+    @SerialName("content") val content: String? = null,
+    @SerialName("created_at") val createdAt: String? = null,
+    @SerialName("id") val id: String? = null,
+    @SerialName("updated_at") val updatedAt: String? = null,
+    @SerialName("url") val url: String? = null
+)
+
