+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowReviewsDto.kt
@@ -1,0 +1,14 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowReviewsDto(
+    @SerialName("id") val id: Int? = null,
+    @SerialName("page") val page: Int? = null,
+    @SerialName("results") val results: List<TvShowReviewDto>? = null,
+    @SerialName("total_pages") val totalPages: Int? = null,
+    @SerialName("total_results") val totalResults: Int? = null
+)
+
