+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/models/TvShowSimilarsDto.kt
@@ -1,0 +1,12 @@
+package com.datasource.remote.tvShow.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TvShowSimilarsDto(
+    @SerialName("page") val page: Int? = null,
+    @SerialName("results") val tvShowSimilarDto: List<TvShowSimilarDto>? = null,
+    @SerialName("total_pages") val totalPages: Int? = null,
+    @SerialName("total_results") val totalResults: Int? = null
+)
