+++ datasource/remote/tvShow/src/main/java/com/datasource/remote/tvShow/utils/NetworkErrorMapper.kt
@@ -1,0 +1,16 @@
+package com.datasource.remote.tvShow.utils
+
+import io.ktor.client.statement.HttpResponse
+
+object NetworkErrorMapper {
+    fun toNetworkException(response: HttpResponse): NetworkException {
+        return when (response.status.value) {
+            401 -> NetworkException.UnauthorizedException("Unauthorized")
+            408 -> NetworkException.RequestTimeoutException("Request Timeout")
+            in 500..599 -> NetworkException.ServerException("Server Error")
+            else -> NetworkException.UnknownException("Unknown network error with status: ${response.status.value}")
+        }
+    }
+}
+
+
